Add tfidf support to creating an FC from HTML.
This also removes support for materializing an FC when running a
search using a non-existent query. I claim that this shouldn't happen
in correct code. Moreover, it can never be right because the backend
cannot scrape pages reliably. The HTML *must* be sent by the user
explicitly.

Start working on a test for it too...