Use a cleverer "already labeled" filtering function.
The default filtering function checks to see if a Unit of Retrieval
already has a direct label with the query. If it does, that UoR is
filtered out of the result set.

In SortingDesk, this process is a little more complicated because
searches happen at the subtopic level. Namely, a result should not
be shown if it has already been labeled with something else inside
the current subfolder. This condition cannot be detected using only
directly connected labels. Instead, we need to find the parent
subfolder and exclude all items in that subfolder from the result set.

Another approach here would be using the subtopic connected
component, but this requires that subtopic ids get passed to filtering
functions, which hasn't happened yet.

(If we do a standard connected component, then we will end up filtering
out too much because it will included everything at the topic level.)