Fix bug related to mounting.
When a Bottle application is mounted, the dependency injectors we use
are applied on the mounted application---but the application returned
from `get_application` is the *root* application, which won't have its
injectors applied to the mounted application.

This is a nasty bug caused by the very poor constructor interface of
dossier.web. For now, we cheat and reach into dossier.web's private
global namespace to get the mounted application.