Beta version of keyword extractor. Open query will now run queries that it learns from training a classifer on your entity of interest.
The folder is the "Name" of the entity. More generally, it's the
generic/obvious keyword search you would have done without knowing
anything about the entity. The name of the folder here is actually
important, because the system uses it as the base query.

The subfolders are different confusable versions of that entity. The
names here aren't used for anything (though they could be, if you
wanted to add additional queries).

If you click Open Query on a subfolder, the system will train a
classifier for for that entity. It uses the bookmarked documents
within that subfolder as positive examples, and it uses all the
bookmarked documents in the other sibling subfolders as negative
examples. It uses the feature GPE "Geo-Political Entity" which seemed
to have pretty good keywords (at least in this case ... I haven't
really explored others because in order for it to work you need to
create a properly annotated set ...)

0. the original name.

1. the original name with the most positive predictive keyword.

2. the original name with minus the least positive predictive keyword.

3. the original name with minus the most predictive keyword for the
negative class.

4. the original name with plus the least predictive keyword for the
negative class.